installer/resources/src/requirements.txt
boto==2.49.0
boto3==1.35.29
botocore==1.35.29
aws-cdk-lib==2.160.0
colored==2.2.4
requests==2.32.3
requests-aws4auth==1.3.1
PyYAML==6.0.2
rich==13.8.1
jinja2==3.1.4
